improve semantics
